+++ src/Data/Aeson.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE NoImplicitPrelude #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 -- |
 -- Module:      Data.Aeson
 -- Copyright:   (c) 2011-2016 Bryan O'Sullivan
@@ -34,6 +35,12 @@ module Data.Aeson
 
     -- ** Direct encoding
     -- $encoding
+
+    -- * Remarks on specific encodings
+    -- ** Time
+    -- $time
+
+    -- * Main encoding and decoding functions
       decode
     , decode'
     , eitherDecode
@@ -49,6 +56,12 @@ module Data.Aeson
     , eitherDecodeFileStrict
     , eitherDecodeStrict'
     , eitherDecodeFileStrict'
+    -- ** Exception throwing variants
+    , AesonException (..)
+    , throwDecode
+    , throwDecodeStrict
+    , throwDecode'
+    , throwDecodeStrict'
     -- * Core JSON types
     , Value(..)
     , ErrorResp(..)
@@ -57,6 +70,7 @@ module Data.Aeson
     , fromEncoding
     , Array
     , Object
+    , Key
     -- * Convenience types
     , DotNetTime(..)
     -- * Type conversion
@@ -150,6 +164,8 @@ module Data.Aeson
 
 import Prelude.Compat
 
+import Control.Exception (Exception (..))
+import Control.Monad.Catch (MonadThrow (..))
 import Data.Aeson.Types.FromJSON (ifromJSON, parseIndexedJSON)
 import Data.Aeson.Encoding (encodingToLazyByteString)
 import Data.Aeson.Parser.Internal (decodeWith, decodeStrictWith, eitherDecodeWith, eitherDecodeStrictWith, jsonEOF, json, jsonEOF', json', addFieldNameToErrorResp)
@@ -157,6 +173,9 @@ import Data.Aeson.Types
 import Data.Aeson.Types.Internal (IResult(..), ErrorResp(..) , ErrorType(..), getFieldName)
 import qualified Data.ByteString as B
 import qualified Data.ByteString.Lazy as L
+
+-- $setup
+-- >>> :set -XOverloadedStrings
 
 -- | Efficiently serialize a JSON value as a lazy 'L.ByteString'.
 --
@@ -281,6 +300,62 @@ eitherDecodeFileStrict' :: (FromJSON a) => FilePath -> IO (Either String a)
 eitherDecodeFileStrict' =
   fmap eitherDecodeStrict' . B.readFile
 {-# INLINE eitherDecodeFileStrict' #-}
+
+throwFormatError :: MonadThrow m => Either (JSONPath, String) a -> m a
+throwFormatError = either (throwM . AesonException . uncurry formatError) return
+{-# INLINE throwFormatError #-}
+
+-- | Like 'decode' but throws an 'AesonException' when decoding fails.
+--
+-- >>> throwDecode "42" :: Maybe Int
+-- Just 42
+--
+-- >>> throwDecode "42" :: IO Int
+-- 42
+--
+-- >>> throwDecode "true" :: IO Int
+-- ...Exception: AesonException...
+--
+-- @since 2.1.2.0
+--
+throwDecode :: forall a m. (FromJSON a, MonadThrow m) => L.ByteString -> m a
+throwDecode = throwFormatError . eitherDecodeWith jsonEOF ifromJSON
+{-# INLINE throwDecode #-}
+
+-- | Like 'decodeStrict' but throws an 'AesonException' when decoding fails.
+--
+-- @since 2.1.2.0
+--
+throwDecodeStrict :: forall a m. (FromJSON a, MonadThrow m) => B.ByteString -> m a
+throwDecodeStrict =
+  throwFormatError . eitherDecodeStrictWith jsonEOF ifromJSON
+{-# INLINE throwDecodeStrict #-}
+
+-- | Like 'decode'' but throws an 'AesonException' when decoding fails.
+--
+-- @since 2.1.2.0
+--
+throwDecode' :: forall a m. (FromJSON a, MonadThrow m) => L.ByteString -> m a
+throwDecode' = throwFormatError . eitherDecodeWith jsonEOF' ifromJSON
+{-# INLINE throwDecode' #-}
+
+-- | Like 'decodeStrict'' but throws an 'AesonException' when decoding fails.
+--
+-- @since 2.1.2.0
+--
+throwDecodeStrict' :: forall a m. (FromJSON a, MonadThrow m) => B.ByteString -> m a
+throwDecodeStrict' =
+  throwFormatError . eitherDecodeStrictWith jsonEOF' ifromJSON
+{-# INLINE throwDecodeStrict' #-}
+
+-- | Exception thrown by 'throwDecode' and variants.
+--
+-- @since 2.1.2.0
+newtype AesonException = AesonException String
+  deriving (Show)
+
+instance Exception AesonException where
+    displayException (AesonException str) = "aeson: " ++ str
 
 -- $use
 --
@@ -530,3 +605,36 @@ eitherDecodeFileStrict' =
 -- > > import Data.Sequence as Seq
 -- > > encode (Seq.fromList [1,2,3])
 -- > "[1,2,3]"
+
+-- $time
+--
+-- This module contains instances of 'ToJSON' and 'FromJSON' for types from
+-- the <https://hackage.haskell.org/package/time time> library.
+--
+-- Those instances encode time as JSON strings in
+-- <https://en.wikipedia.org/wiki/ISO_8601 ISO 8601> formats, with the
+-- following general form for 'Data.Time.Clock.UTCTime' and
+-- 'Data.Time.LocalTime.ZonedTime', while other time types use subsets of those
+-- fields:
+--
+-- > [+,-]YYYY-MM-DDThh:mm[:ss[.sss]]Z
+--
+-- where
+--
+-- - @[+,-]@ is an optional sign, @+@ or @-@.
+-- - @YYYY@ is the year, which must have at least 4 digits to prevent Y2K problems.
+--   Years from @0000@ to @0999@ must thus be zero-padded.
+-- - @MM@ is a two-digit month.
+-- - @DD@ is a two-digit day.
+-- - @T@ is a literal @\'T\'@ character separating the date and the time of
+--   day. It may be a space instead.
+-- - @hh@ is a two-digit hour.
+-- - @mm@ is a two-digit minute.
+-- - @ss@ is a two-digit second.
+-- - @sss@ is a decimal fraction of a second; it may have any nonzero number of digits.
+-- - @Z@ is a time zone; it may be preceded by an optional space.
+--
+-- For more information, see <https://en.wikipedia.org/wiki/ISO_8601 ISO 8601>
+-- <https://hackage.haskell.org/package/time time>,
+-- and <https://hackage.haskell.org/package/attoparsec-iso8601 attoparsec-iso8601>
+-- (where the relevant parsers are defined).