+++ src/Data/Aeson/Types/Internal.hs
@@ -5,7 +5,7 @@
 {-# LANGUAGE NoImplicitPrelude #-}
 {-# LANGUAGE Rank2Types #-}
 {-# LANGUAGE StandaloneDeriving #-}
-{-# LANGUAGE RecordWildCards #-}
+{-# LANGUAGE RecordWildCards,LambdaCase,ScopedTypeVariables #-}
 #if __GLASGOW_HASKELL__ >= 800
 -- a) THQ works on cross-compilers and unregisterised GHCs
 -- b) may make compilation faster as no dynamic loading is ever needed (not sure about this)
@@ -30,6 +30,7 @@ module Data.Aeson.Types.Internal
     (
     -- * Core JSON types
       Value(..)
+    , Key
     , Array
     , emptyArray, isEmptyArray
     , Pair
@@ -45,6 +46,7 @@ module Data.Aeson.Types.Internal
     , ErrorResp(..)
     , ErrorType(..)
     , iparse
+    , iparseEither
     , parse
     , parseEither
     , parseMaybe
@@ -97,16 +99,15 @@ module Data.Aeson.Types.Internal
 import Prelude.Compat
 
 import Control.Applicative (Alternative(..))
-import Control.Arrow (first)
 import Control.DeepSeq (NFData(..))
 import Control.Monad (MonadPlus(..), ap)
+import Control.Monad.Fix (MonadFix (..))
 import Data.Char (isLower, isUpper, toLower, isAlpha, isAlphaNum)
+import Data.Aeson.Key (Key)
 import Data.Data (Data)
 import Data.Foldable (foldl')
-import Data.HashMap.Strict (HashMap)
 import Data.Hashable (Hashable(..))
-import Data.List (intercalate, sortBy)
-import Data.Ord (comparing)
+import Data.List (intercalate)
 import Data.Scientific (Scientific)
 import Data.String (IsString(..))
 import Data.Text (Text, pack, unpack)
@@ -115,35 +116,143 @@ import Data.Time.Format (FormatTime)
 import Data.Typeable (Typeable)
 import Data.Vector (Vector)
 import GHC.Generics (Generic)
+import Data.Aeson.KeyMap (KeyMap)
 import qualified Control.Monad as Monad
 import qualified Control.Monad.Fail as Fail
-import qualified Data.HashMap.Strict as H
-import qualified Data.Scientific as S
 import qualified Data.Vector as V
 import qualified Language.Haskell.TH.Syntax as TH
-import Text.Read (readMaybe)
+import Text.Read (readMaybe,readEither)
+import qualified Data.Aeson.Key as Key
+import qualified Data.Aeson.KeyMap as KM
+import qualified Data.Scientific as Sci
+import qualified Data.Text as T
+import qualified Test.QuickCheck as QC
+import Witherable (ordNub)
 
+import Text.JSON hiding (Error,Result)
+import qualified Text.JSON
 -- | Elements of a JSON path used to describe the location of an
 -- error.
-data JSONPathElement = Key Text
+data JSONPathElement = Key Key
                        -- ^ JSON path element of a key into an object,
                        -- \"object.key\".
                      | Index {-# UNPACK #-} !Int
                        -- ^ JSON path element of an index into an
                        -- array, \"array[index]\".
-                       deriving (Eq, Show, Typeable, Ord, Read)
+                       deriving (Eq, Typeable, Ord)
 type JSONPath = [JSONPathElement]
 
 -- | The internal result of running a 'Parser'.
 data IResult a = IError JSONPath String
                | ISuccess a
-               deriving (Eq, Show, Typeable)
+               deriving (Eq, Typeable)
 
 data ErrorResp = ErrorResp {errorType :: ErrorType, errField :: Maybe String, objectType :: Maybe String, errMessage :: Maybe String , expectedValue :: Maybe String, actualValue :: Maybe String, jPath :: Maybe JSONPath}
-    deriving (Eq, Show, Typeable, Read)
+    deriving (Eq, Typeable)
 
 data ErrorType = MISSING_FIELD | TYPE_MISMATCH | GENERAL
-    deriving (Eq, Show, Typeable, Read)
+    deriving (Eq, Typeable)
+
+maybeFromObj :: JSON a => String -> JSObject JSValue -> Text.JSON.Result (Maybe a)
+maybeFromObj key obj = case lookupObj key obj of
+    Just JSNull -> Ok Nothing
+    Just val -> Just <$> readJSON val
+    Nothing -> Ok Nothing
+
+lookupObj :: String -> JSObject JSValue -> Maybe JSValue
+lookupObj k obj = lookup k (fromJSObject obj)
+
+findObj :: JSON a => String -> JSObject JSValue -> Text.JSON.Result a
+findObj k obj = 
+    case lookupObj k obj of
+        Nothing -> Text.JSON.Error $ "Missing key: " ++ k
+        Just v -> readJSON v
+
+valFromObj :: JSON a => String -> JSObject JSValue -> Text.JSON.Result a
+valFromObj k obj = case lookupObj k obj of
+    Just v -> readJSON v 
+    Nothing -> Text.JSON.Error $ "Required field missing: " ++ k
+
+instance JSON ErrorType where
+    readJSON (JSString s) = case fromJSString s of
+        "MISSING_FIELD" -> Ok MISSING_FIELD
+        "TYPE_MISMATCH" -> Ok TYPE_MISMATCH
+        "GENERAL" -> Ok GENERAL
+        x -> Text.JSON.Error $ "Invalid ErrorType: " ++ x
+    readJSON x = Text.JSON.Error $ "Expected String for ErrorType, got: " ++ show x
+
+    showJSON = JSString . toJSString . \case
+        MISSING_FIELD -> "MISSING_FIELD"
+        TYPE_MISMATCH -> "TYPE_MISMATCH"
+        GENERAL -> "GENERAL"
+
+instance JSON JSONPathElement where
+    readJSON (JSObject obj) = do
+        typ <- Text.JSON.valFromObj "type" obj
+        case typ of
+            "key" -> do
+                val <- Text.JSON.valFromObj "value" obj
+                return $ Key (Key.fromString val)
+            "index" -> do
+                val <- Text.JSON.valFromObj "value" obj
+                return $ Index val
+            _ -> Text.JSON.Error $ "Invalid JSONPathElement type: " ++ typ
+    readJSON x = Text.JSON.Error $ "Expected Object for JSONPathElement, got: " ++ show x
+
+    showJSON (Key k) = makeObj
+        [ ("type", showJSON "key")
+        , ("value", showJSON $ Key.toString k)
+        ]
+    showJSON (Index i) = makeObj
+        [ ("type", showJSON "index")
+        , ("value", showJSON i)
+        ]
+
+instance JSON ErrorResp where
+    readJSON (JSObject obj) = do
+        errorType <- Text.JSON.valFromObj "errorType" obj
+        errField <- maybeFromObj "errField" obj
+        objectType <- maybeFromObj "objectType" obj
+        errMessage <- maybeFromObj "errMessage" obj
+        expectedValue <- maybeFromObj "expectedValue" obj
+        actualValue <- maybeFromObj "actualValue" obj
+        jPath <- case lookupObj "jPath" obj of
+            Just JSNull -> return Nothing
+            Just arr -> Just <$> readJSON arr
+            Nothing -> return Nothing
+        return ErrorResp{..}
+    readJSON x = Text.JSON.Error $ "Expected Object for ErrorResp, got: " ++ show x
+
+    showJSON ErrorResp{..} = makeObj
+        [ ("errorType", showJSON errorType)
+        , ("errField", maybe JSNull showJSON errField)
+        , ("objectType", maybe JSNull showJSON objectType)
+        , ("errMessage", maybe JSNull showJSON errMessage)
+        , ("expectedValue", maybe JSNull showJSON expectedValue)
+        , ("actualValue", maybe JSNull showJSON actualValue)
+        , ("jPath", maybe JSNull showJSON jPath)
+        ]
+
+instance Show ErrorResp where
+    show = encode
+
+instance Read ErrorResp where
+    readsPrec _ input = case decode input of
+        Ok val -> [(val, "")]
+        --NOTE: not throwing error since , readMaybe or readEither are also throwing exception
+        Text.JSON.Error msg -> []--error msg
+
+example :: ErrorResp
+example = ErrorResp {
+    errorType = TYPE_MISMATCH,
+    errField = Just "age",
+    objectType = Just "User",
+    errMessage = Just "Invalid type for age",
+    expectedValue = Just "number",
+    actualValue = Just "string",
+    jPath = Just [Key (Key.fromString "user"), Index 0, Key (Key.fromString "profile"), Key (Key.fromString "age")]
+}
+
 -- | The result of running a 'Parser'.
 data Result a = Error String
               | Success a
@@ -319,6 +428,19 @@ instance Monad.Monad Parser where
     {-# INLINE fail #-}
 #endif
 
+-- |
+--
+-- @since 2.1.0.0
+instance MonadFix Parser where
+    mfix f = Parser $ \path kf ks -> let x = runParser (f (fromISuccess x)) path IError ISuccess in
+        case x of
+            IError p e -> kf p e
+            ISuccess y -> ks y
+      where
+        fromISuccess :: IResult a -> a
+        fromISuccess (ISuccess x)      = x
+        fromISuccess (IError path msg) = error $ "mfix @Aeson.Parser: " ++ formatPath path ++ ": " ++ msg
+
 instance Fail.MonadFail Parser where
     fail msg = Parser $ \path kf _ks -> kf (reverse path) msg
     {-# INLINE fail #-}
@@ -368,7 +490,7 @@ apP d e = do
 {-# INLINE apP #-}
 
 -- | A JSON \"object\" (key\/value map).
-type Object = HashMap Text Value
+type Object = KeyMap Value
 
 -- | A JSON \"array\" (sequence).
 type Array = Vector Value
@@ -402,8 +524,95 @@ instance Show Value where
         $ showString "Array " . showsPrec 11 xs
     showsPrec d (Object xs) = showParen (d > 10)
         $ showString "Object (fromList "
-        . showsPrec 11 (sortBy (comparing fst) (H.toList xs))
+        . showsPrec 11 (KM.toAscList xs)
         . showChar ')'
+
+-- | @since 2.0.3.0
+instance QC.Arbitrary Value where
+    arbitrary = QC.sized arbValue
+
+    shrink = ordNub . go where
+        go Null       = []
+        go (Bool b)   = Null : map Bool (QC.shrink b)
+        go (String x) = Null : map (String . T.pack) (QC.shrink (T.unpack x))
+        go (Number x) = Null : map Number (shrScientific x)
+        go (Array x)  = Null : V.toList x ++ map (Array . V.fromList) (QC.liftShrink go (V.toList x))
+        go (Object x) = Null : KM.elems x ++ map (Object . KM.fromList) (QC.liftShrink (QC.liftShrink go) (KM.toList x))
+
+-- | @since 2.0.3.0
+instance QC.CoArbitrary Value where
+    coarbitrary Null       = QC.variant (0 :: Int)
+    coarbitrary (Bool b)   = QC.variant (1 :: Int) . QC.coarbitrary b
+    coarbitrary (String x) = QC.variant (2 :: Int) . QC.coarbitrary (T.unpack x)
+    coarbitrary (Number x) = QC.variant (3 :: Int) . QC.coarbitrary (Sci.coefficient x) . QC.coarbitrary (Sci.base10Exponent x)
+    coarbitrary (Array x)  = QC.variant (4 :: Int) . QC.coarbitrary (V.toList x)
+    coarbitrary (Object x) = QC.variant (5 :: Int) . QC.coarbitrary (KM.toList x)
+
+-- | @since 2.0.3.0
+instance QC.Function Value where
+    function = QC.functionMap fwd bwd where
+        fwd :: Value -> RepValue
+        fwd Null       = Left Nothing
+        fwd (Bool b)   = Left (Just b)
+        fwd (String x) = Right (Left (Left (T.unpack x)))
+        fwd (Number x) = Right (Left (Right (Sci.coefficient x, Sci.base10Exponent x)))
+        fwd (Array x)  = Right (Right (Left (V.toList x)))
+        fwd (Object x) = Right (Right (Right (KM.toList x)))
+
+        bwd :: RepValue -> Value
+        bwd (Left Nothing)                = Null
+        bwd (Left (Just b))               = Bool b
+        bwd (Right (Left (Left x)))       = String (T.pack x)
+        bwd (Right (Left (Right (x, y)))) = Number (Sci.scientific x y)
+        bwd (Right (Right (Left x)))      = Array (V.fromList x)
+        bwd (Right (Right (Right x)))     = Object (KM.fromList x)
+
+-- Used to implement QC.Function Value instance
+type RepValue
+    = Either (Maybe Bool) (Either (Either String (Integer, Int)) (Either [Value] [(Key, Value)]))
+
+arbValue :: Int -> QC.Gen Value
+arbValue n
+    | n <= 0 = QC.oneof
+        [ pure Null
+        , Bool <$> QC.arbitrary
+        , String <$> arbText
+        , Number <$> arbScientific
+        ]
+
+    | otherwise = QC.oneof
+        [ Object <$> arbObject n
+        , Array <$> arbArray  n
+        ]
+
+arbText :: QC.Gen Text
+arbText = T.pack <$> QC.arbitrary
+
+arbScientific :: QC.Gen Scientific
+arbScientific = Sci.scientific <$> QC.arbitrary <*> QC.arbitrary
+
+shrScientific :: Scientific -> [Scientific]
+shrScientific s = map (uncurry Sci.scientific) $
+    QC.shrink (Sci.coefficient s, Sci.base10Exponent s) 
+
+arbObject :: Int -> QC.Gen Object
+arbObject n = do
+    p <- arbPartition (n - 1)
+    KM.fromList <$> traverse (\m -> (,) <$> QC.arbitrary <*> arbValue m) p
+
+arbArray :: Int -> QC.Gen Array
+arbArray n = do
+    p <- arbPartition (n - 1)
+    V.fromList <$> traverse arbValue p
+
+arbPartition :: Int -> QC.Gen [Int]
+arbPartition k = case compare k 1 of
+    LT -> pure []
+    EQ -> pure [1]
+    GT -> do
+        first <- QC.chooseInt (1, k)
+        rest <- arbPartition $ k - first
+        QC.shuffle (first : rest)
 
 -- |
 --
@@ -453,20 +662,17 @@ hashValue s Null         = s `hashWithSalt` (5::Int)
 instance Hashable Value where
     hashWithSalt = hashValue
 
--- @since 0.11.0.0
+-- | @since 0.11.0.0
 instance TH.Lift Value where
-    lift Null = [| Null |]
-    lift (Bool b) = [| Bool b |]
-    lift (Number n) = [| Number (S.scientific c e) |]
-      where
-        c = S.coefficient n
-        e = S.base10Exponent n
+    lift Null       = [| Null |]
+    lift (Bool b)   = [| Bool b |]
+    lift (Number n) = [| Number n |]
     lift (String t) = [| String (pack s) |]
       where s = unpack t
-    lift (Array a) = [| Array (V.fromList a') |]
+    lift (Array a)  = [| Array (V.fromList a') |]
       where a' = V.toList a
-    lift (Object o) = [| Object (H.fromList . map (first pack) $ o') |]
-      where o' = map (first unpack) . H.toList $ o
+    lift (Object o) = [| Object o |]
+
 #if MIN_VERSION_template_haskell(2,17,0)
     liftTyped = TH.unsafeCodeCoerce . TH.lift
 #elif MIN_VERSION_template_haskell(2,16,0)
@@ -485,11 +691,11 @@ isEmptyArray _ = False
 
 -- | The empty object.
 emptyObject :: Value
-emptyObject = Object H.empty
+emptyObject = Object KM.empty
 
 -- | Run a 'Parser'.
 parse :: (a -> Parser b) -> a -> Result b
-parse m v = runParser (m v) [] (const Error) Success
+parse m v = runParser (m v) [] (const Data.Aeson.Types.Internal.Error) Success
 {-# INLINE parse #-}
 
 -- | Run a 'Parser'.
@@ -508,6 +714,14 @@ parseEither :: (a -> Parser b) -> a -> Either String b
 parseEither m v = runParser (m v) [] onError Right
   where onError path msg = Left (addFieldNameToErrorResp path msg)
 {-# INLINE parseEither #-}
+
+-- | Run a 'Parser' with an 'Either' result type.
+-- If the parse fails, the 'Left' payload will contain an error message and a json path to failed element.
+--
+-- @since 2.1.0.0
+iparseEither :: (a -> Parser b) -> a -> Either (JSONPath, String) b
+iparseEither m v = runParser (m v) [] (\path msg -> Left (path, msg)) Right
+{-# INLINE iparseEither #-}
 
 -- | Annotate an error message with a
 -- <http://goessner.net/articles/JsonPath/ JSONPath> error location.
@@ -529,11 +743,11 @@ formatRelativePath path = format "" path
     format pfx (Index idx:parts) = format (pfx ++ "[" ++ show idx ++ "]") parts
     format pfx (Key key:parts)   = format (pfx ++ formatKey key) parts
 
-    formatKey :: Text -> String
+    formatKey :: Key -> String
     formatKey key
        | isIdentifierKey strKey = "." ++ strKey
        | otherwise              = "['" ++ escapeKey strKey ++ "']"
-      where strKey = unpack key
+      where strKey = Key.toString key
 
     isIdentifierKey :: String -> Bool
     isIdentifierKey []     = False
@@ -554,7 +768,7 @@ getFieldName path =
         format :: JSONPath -> Maybe String
         format []                = Nothing
         format (Index idx:parts) = format parts
-        format (Key key:_)   = Just $ formatKey key
+        format (Key key:_)   = Just $ formatKey (Key.toText key)
 
         formatKey :: Text -> String
         formatKey key
@@ -613,12 +827,12 @@ missingFieldErr :: Maybe String -> String -> String
 missingFieldErr objectType field = show $
             defaultErrorObject {errorType = MISSING_FIELD, errField = Just field, objectType = objectType}
 -- | A key\/value pair for an 'Object'.
-type Pair = (Text, Value)
+type Pair = (Key, Value)
 
 -- | Create a 'Value' from a list of name\/value 'Pair's.  If duplicate
--- keys arise, earlier keys and their associated values win.
+-- keys arise, later keys and their associated values win.
 object :: [Pair] -> Value
-object = Object . H.fromList
+object = Object . KM.fromList
 {-# INLINE object #-}
 
 -- | Add JSON Path context to a parser
@@ -915,8 +1129,8 @@ camelTo c = lastWasCap True
 
 -- | Better version of 'camelTo'. Example where it works better:
 --
---   > camelTo '_' 'CamelAPICase' == "camel_apicase"
---   > camelTo2 '_' 'CamelAPICase' == "camel_api_case"
+--   > camelTo '_' "CamelAPICase" == "camel_apicase"
+--   > camelTo2 '_' "CamelAPICase" == "camel_api_case"
 camelTo2 :: Char -> String -> String
 camelTo2 c = map toLower . go2 . go1
     where go1 "" = ""